requirements.txt
requests
python-whois
tldextract
beautifulsoup4
pandas
scikit-learn
joblib

